WIP refactor to allow inheritance + demo
