+++ package.json
@@ -89,6 +89,7 @@
     "@angular/compiler": "^5.0.2",
     "@angular/compiler-cli": "^5.0.2",
     "@angular/core": "^5.0.2",
+    "@angular/flex-layout": "^2.0.0-beta.10-4905443",
     "@angular/forms": "^5.0.2",
     "@angular/http": "^5.0.2",
     "@angular/material": "^5.0.0",
