+++ src/demo/app/home/home.component.html
@@ -27,7 +27,8 @@
 
   <h3>Features</h3>
   <ul>
-    <li>Supports IPv4 and IPv6</li>
+    <li>Ability to customize the internal <b>input</b> controls <button mat-button color="primary" (click)="openCustomComponentDemo()">See Demo</button></li>
+    <li>Supports IPv4, IPv6 and MAC address</li>
     <li>Control if user can type invalid characters</li>
     <li>Separator behaves like TAB (e.g: in IPv4 pressing dot (.) is like pressing TAB)</li>
     <li>Support paste (single value or whole address)</li>
