+++ src/demo/app/home/home.component.ts
@@ -2,6 +2,7 @@ import { Component, ViewContainerRef } from '@angular/core';
 import { MatDialog, MatDialogRef } from '@angular/material';
 
 import { DemoDialogComponent } from './demo-dialog/demo-dialog.component';
+import { CustomIpComponentDialogComponent } from './custom-ip-component/custom-ip-component-dialog.component';
 
 @Component({
   selector: 'home',
@@ -10,7 +11,7 @@ import { DemoDialogComponent } from './demo-dialog/demo-dialog.component';
 export class HomeComponent {
   ip: string = '192.333.0.0';
 
-  dialogRef: MatDialogRef<DemoDialogComponent>;
+  dialogRef: MatDialogRef<DemoDialogComponent | CustomIpComponentDialogComponent>;
 
   constructor(public dialog: MatDialog,
               public viewContainerRef: ViewContainerRef) {
@@ -19,6 +20,13 @@ export class HomeComponent {
 
   openDemo() {
     this.dialogRef = this.dialog.open(DemoDialogComponent, {
+      viewContainerRef: this.viewContainerRef,
+      role: 'dialog'
+    });
+  }
+
+  openCustomComponentDemo() {
+    this.dialogRef = this.dialog.open(CustomIpComponentDialogComponent, {
       viewContainerRef: this.viewContainerRef,
       role: 'dialog'
     });
