+++ src/demo/app/home/index.ts
@@ -1,2 +1,4 @@
 export { HomeComponent } from './home.component';
 export { DemoDialogComponent } from './demo-dialog/demo-dialog.component';
+export { CustomIpComponentDialogComponent } from './custom-ip-component/custom-ip-component-dialog.component';
+export { CustomNgxIpComponent } from './custom-ip-component/custom-ngx-ip.component';
