+++ src/ngx-ip/src/index.ts
@@ -1,2 +1,9 @@
 export { NgxIpModule } from './ngx-ip.module';
-export { VALIDATION_TYPE, ADDRESS_MODE_TYPE, COPY_MODE_TYPE } from './ngx-ip.component';
+export { NgxIpComponent } from './ngx-ip.component';
+export {
+  VALIDATION_TYPE,
+  ADDRESS_MODE_TYPE,
+  COPY_MODE_TYPE,
+  COPY_METHOD,
+  NgxIpBase
+} from './ngx-ip.base';
